rollback move on inedx
